pyproject.toml
[project]
name = "chat2bet"
version = "0.1.0"
description = ""
authors = [
    {name = "tomasz-wlodarczyk-pawatech",email = "tomasz.wlodarczyk@pawatech.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
beautifulsoup4 = "*"
langchain-community = "*"
tiktoken = "*"
langchainhub = "*"
langchain = "*"
langgraph = "*"
tavily-python = "*"
langchain-openai = "*"
python-dotenv = "*"
black = "*"
isort = "*"
pytest = "*"
langchain-chroma = "*"
fastapi = "*"
langchain-tavily = "^0.1.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
